[~] change file arch [~] remove warnings

+++ oss_lib/src/printf/my_putnbr_base.c
@@ -10,7 +10,7 @@
 int get_display_length(unsigned long long int nbr, char const *base)
 {
     int i = 0;
-    int len = 0;
+    unsigned long long int len = 0;
 
     for (; base[len] != '\0'; len++);
     while (nbr >= len){
@@ -23,7 +23,7 @@ int get_display_length(unsigned long long int nbr, char const *base)
 void my_putnbr_base(unsigned long long int nbr, char const *base,
     int disp_len, pf_conf_t *conf)
 {
-    int len = 0;
+    unsigned long long int len = 0;
     int i = 0;
     int checker;
     char mem_display[disp_len];
